[Refactor] Extract messageFor and networkSendSystem
